% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatroom.R
\name{create_aPrivateRoomUnderGroup}
\alias{create_aPrivateRoomUnderGroup}
\title{Create a room under group}
\usage{
create_aPrivateRoomUnderGroup(
  groupId,
  name,
  topic = "",
  securityType = "PRIVATE"
)
}
\arguments{
\item{groupId}{A character of groupId}

\item{name}{A character determing the prefix of room name attached by its sha1 10-digit code}

\item{topic}{A character, default=""}

\item{securityType}{A character, default="PRIVATE". The other option is "PUBLIC"}
}
\value{

}
\description{
Create a room under group
}
\examples{
none.
}
