% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitter2github.R
\name{extractPrivateGitterChatMessage}
\alias{extractPrivateGitterChatMessage}
\title{Extract private gitter message and output its html content to the clipboard}
\usage{
extractPrivateGitterChatMessage()
}
\value{

}
\description{
Copy the first two line of a chat (first line is like "kkqq123123123 @kkqq123123123 12:23", the second line sentence will be used to define the starting location of message extraction.)
}
\examples{
none.
}
