% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_issueWithGetCommentsAttached}
\alias{get_issueWithGetCommentsAttached}
\title{Get issue with get_comments method attached}
\usage{
get_issueWithGetCommentsAttached(owner, repo, issue_number, ...)
}
\arguments{
\item{owner}{A character of owner}

\item{repo}{A character of repo name}

\item{issue_number}{A character of issue number}

\item{...}{other query parameters}
}
\value{
A list of response where element comments_url has get_comments method attached.
}
\description{
Get issue with get_comments method attached
}
\examples{
none
}
