% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubMonitor.R
\name{getUserEvents_afterTime}
\alias{getUserEvents_afterTime}
\title{Get user events whose created at time is after certain time}
\usage{
getUserEvents_afterTime(username, startTime)
}
\arguments{
\item{username}{A character}

\item{startTime}{A Date/Time}
}
\value{
A list of possible events that fit the criterion
}
\description{
Get user events whose created at time is after certain time
}
