% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{githubService}
\alias{githubService}
\title{Create a Github service}
\usage{
githubService()
}
\value{
a list of service
}
\description{
Create a Github service
}
\examples{
gs <- githubService()
gs$get_userInfo
gs$get_allRepos(username = "tpemartin")
gs$get_oneRepoInfo(owner = "tpemartin", repo="109-1-inclass-practice")
gs$get_repoForks(owner = "tpemartin", repo="109-1-inclass-practice")
}
