% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitter.R
\name{gitterService}
\alias{gitterService}
\title{Create gitter service}
\usage{
gitterService()
}
\value{

}
\description{
The service contains list_groups/list_groupsAsDataframe, list_roomsUnderAGroup/list_roomsUnderAGroupAsDataFrame(groupId), list_usersInARoom/list_usersInARoomAsDataFrame(roomId)
}
\examples{
gs <- gitterService()
}
