% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{rootEndpoints}
\alias{rootEndpoints}
\title{Define the root endpoint of api service}
\usage{
rootEndpoints(platformName)
}
\arguments{
\item{platformName}{A character of "github", "gitter"}
}
\value{
endpoint url
}
\description{
Define the root endpoint of api service
}
\examples{
none
}
